Add JWT token to authorized users
To be more secured we going to use JesonWebToken(JWT) and decode it on
frontend site.
Token has all needed date and it refresh after 90 days.
There is link to backend connected to users but it only allow POST
method.
Giving permisions is provided by admin site.
In feature we must change logic to update user data!