HotFix construtor in Places
Some fields in models of Places are set as default None and thats
means not what I think. Constructor of Places trows error every time
when we try post some data without all fields given in places. Now it
works as it should when no needed data is not given it is null.